backend/requirements.txt
fastapi
uvicorn
livekit
livekit-agents
faster-whisper
pydantic
python-dotenv
python-multipart
aiohttp
